complete classes computing local / transition costs, added testcases, starting to implement DP

+++ src/samplingemissions.cpp
@@ -1,6 +1,7 @@
 #include <map>
 #include <cmath>
 #include <cassert>
+#include <algorithm>
 #include "samplingemissions.hpp"
 
 using namespace std;
@@ -19,11 +20,11 @@ SamplingEmissions::SamplingEmissions(shared_ptr<UniqueKmers> unique_kmers) {
 			assert(this->allele_penalties[it->first] < 25);
 		} else {
 			// Note: this value is based on the max number of unique kmers (which is 300)
-			this->allele_penalties[it->second] = 25;
+			this->allele_penalties[it->second] = 25.0;
 		}
 	}
 }
 
-float SamplingEmissions::get_emission_cost(unsigned char allele_id) {
+unsigned int SamplingEmissions::get_emission_cost(unsigned char allele_id) {
 	return this->allele_penalties[allele_id];
 }
